Steady the empty state's animation and give the page its own class
The attention pulse was cancelled wherever it happened to stand, a few
percent up, while both the reveal and the next pulse start from 1f — so
reopening the app from the background snapped the call to action back
down. That trip also replayed the entire reveal on a page already in
front of the user, and toggled the pill's hardware layer off and on
around it, which showed as the button flicking.

The page now reveals only when it actually appears, the pulse is bounded
to three breaths and hands its layer back when it ends, and stopping it
always restores the scale. A page left open used to redraw at the
display's refresh rate for as long as it was up (609 frames over ten idle
seconds); it now renders nothing once the pulse is over.

Everything the page owns moves out of PlayerActivity into EmptyState,
which is 262 lines the activity no longer carries. isReducedMotion joins
Utils, since the activity needs it for its own marquee decision.
